requirements.txt
pytesseract
pdf2image
Pillow
opencv-python-headless
ocrmypdf
pdfplumber